requirements.txt
python-dotenv
nvdlib
discord
argparse